+++ starknet-rpc-test/src/lib.rs
@@ -179,7 +179,7 @@ impl MadaraClient {
             results.push(result);
         }
 
-        self.do_create_block(true, false).await?;
+        self.do_create_block(true, true).await?;
         Ok(results)
     }